requirements.txt
google-auth==2.17.3
requests==2.28.2
urllib3==1.26.15